Added new features. Add file feature is broken
Add file = not working for files >= 40 mb

+++ src/server/global.cpp
@@ -10,7 +10,7 @@ net::Server<char> *server = nullptr;
 
 bool DEBUG_FLAG = false;
 
-long long unsigned int TOTAL_BYTES_RECV = 0;
+double TOTAL_BYTES_RECV = 0;
 
 bool findString(const char haystack[], const char needle[])
 {
